+++ src/onegov/feriennet/forms/occasion.py
@@ -211,6 +211,12 @@ class OccasionForm(Form):
             self.dates.errors = [_("Must specify at least one date")]
             return False
 
+    def ensure_min_max_age(self):
+        if self.min_age.data > self.max_age.data:
+            self.min_age.errors = [
+                _("Minimum Age must be lower than maximum age")]
+            return False
+
     def ensure_safe_period_change(self):
         # the period may only be changed if there are no booking associated
         # with the occasion, otherwise this is unsafe and results in