+++ src/onegov/org/mail.py
@@ -44,10 +44,21 @@ def send_marketing_html_mail(*args, **kwargs):
 
 def send_ticket_mail(request, template, subject, receivers, ticket,
                      content=None, force=False, **kwargs):
-
+    org = request.app.org
     if not force:
 
+        if org.mute_all_tickets:
+            return
+
         if ticket.muted:
+            return
+
+        skip_handler_codes = org.tickets_skip_opening_email or []
+        opened = ticket.state == 'open'
+        if opened and ticket.handler_code in skip_handler_codes:
+            return
+
+        if opened and request.auto_accept(ticket):
             return
 
         if request.current_username in receivers: