+++ src/onegov/org/models/ticket.py
@@ -280,6 +280,8 @@ class ReservationHandler(Handler):
 
     @cached_property
     def resource(self):
+        if self.deleted:
+            return None
         query = self.session.query(Resource)
         query = query.filter(Resource.id == self.reservations[0].resource)
 
@@ -318,6 +320,8 @@ class ReservationHandler(Handler):
     @property
     def email(self):
         # the e-mail is the same over all reservations
+        if self.deleted:
+            return self.ticket.snapshot.get('email')
         return self.reservations[0].email
 
     @property
@@ -519,8 +523,18 @@ class EventSubmissionHandler(Handler):
         return self.event is None
 
     @cached_property
+    def source(self):
+        # values stored only when importing with cli
+        return self.data.get('source')
+
+    @cached_property
+    def import_user(self):
+        # values stored only when importing with cli
+        return self.data.get('user')
+
+    @cached_property
     def email(self):
-        return self.event.meta.get('submitter_email')
+        return self.event and self.event.meta.get('submitter_email')
 
     @property
     def title(self):
@@ -599,6 +613,31 @@ class EventSubmissionHandler(Handler):
                         )
                     )
                 )
+            ) if not self.event.source else (
+                Link(
+                    text=_("Withdraw event"),
+                    url=layout.csrf_protected_url(request.link(
+                        self.event, name='withdraw')),
+                    attrs={'class': ('delete-link')},
+                    traits=(
+                        Confirm(
+                            _("Do you really want to withdraw this event?"),
+                            _("You can re-publish an imported event later."),
+                            _("Withdraw event"),
+                            _("Cancel")
+                        ),
+                        Intercooler(
+                            request_method='POST',
+                            redirect_after=request.link(self.ticket)
+                        )
+                    )
+                ) if not self.event.state == 'withdrawn'
+                else Link(
+                    text=_("Re-publish event"),
+                    url=request.return_here(
+                        request.link(self.event, 'publish')),
+                    attrs={'class': 'accept-link'}
+                ),
             ), right_side=False))
 
         return links