+++ tests/onegov/event/test_collections.py
@@ -225,6 +225,9 @@ def test_occurrence_collection_query(session):
     assert query(outdated=True, locations=['Squirrel']).count() == 5
     assert query(outdated=True, locations=['Squirrel Park']).count() == 5
     assert query(outdated=True, locations=['Squirrel', 'Park']).count() == 5
+    # test random user input
+    assert query(
+        outdated=True, locations=['Squirrel Park\'\',,"").")']).count() == 5
     assert query(outdated=True, locations=['Center']).count() == 1
     assert query(outdated=True, locations=['squirrel', 'park']).count() == 0
     assert query(outdated=True, locations=[]).count() == 5
@@ -769,7 +772,7 @@ def test_as_ical(session):
 def test_from_import(session):
     events = EventCollection(session)
 
-    assert events.from_import([
+    added, updated, purged = events.from_import([
         EventImportItem(
             event=Event(
                 state='initiated',
@@ -818,7 +821,8 @@ def test_from_import(session):
             pdf=None,
             pdf_filename=None,
         )
-    ]) == (2, 0, 0)
+    ])
+    assert (len(added), len(updated), len(purged)) == (2, 0, 0)
 
     def items():
         yield EventImportItem(
@@ -846,7 +850,8 @@ def test_from_import(session):
             pdf_filename=None,
         )
 
-    assert events.from_import(items()) == (1, 0, 0)
+    added, updated, purged = events.from_import(items())
+    assert (len(added), len(updated), len(purged)) == (1, 0, 0)
 
     # Already imported
     assert events.from_import([
@@ -874,10 +879,10 @@ def test_from_import(session):
             pdf=None,
             pdf_filename=None,
         )
-    ]) == (0, 0, 0)
+    ]) == ([], [], [])
 
     # Update and purge
-    assert events.from_import([
+    a, u, p = events.from_import([
         EventImportItem(
             event=Event(
                 state='initiated',
@@ -902,11 +907,12 @@ def test_from_import(session):
             pdf=None,
             pdf_filename=None,
         )
-    ], 'import-1') == (0, 1, 1)
+    ], 'import-1')
+    assert (len(a), len(u), len(p)) == (0, 1, 1)
     assert events.subset_count == 2
 
     # Don't purge
-    assert events.from_import(['import-1-A'], 'import-1') == (0, 0, 0)
+    assert events.from_import(['import-1-A'], 'import-1') == ([], [], [])
     assert events.subset_count == 2
 
     # Withdraw
@@ -936,7 +942,7 @@ def test_from_import(session):
             pdf=None,
             pdf_filename=None,
         )
-    ]) == (0, 0, 0)
+    ]) == ([], [], [])
     assert events.by_name('title-c').state == 'withdrawn'